requirements.txt
# Face Sorter Requirements
insightface
scikit-learn
pillow
pillow-heif
tqdm
numpy
matplotlib
onnxruntime
